+++ backend/store/store.go
@@ -2,8 +2,10 @@ package store
 
 import (
 	"context"
+	"crypto/tls"
 	"errors"
 
+	"github.com/coreos/etcd/clientv3"
 	"github.com/sensu/sensu-go/types"
 )
 
@@ -289,7 +291,7 @@ type HandlerStore interface {
 
 // HealthStore provides methods for cluster health
 type HealthStore interface {
-	GetClusterHealth(ctx context.Context) *types.HealthResponse
+	GetClusterHealth(ctx context.Context, cluster clientv3.Cluster, etcdClientTLSConfig *tls.Config) *types.HealthResponse
 }
 
 // KeepaliveStore provides methods for managing entities keepalives